fix styles

+++ src/utils/functions/formatMessageTime.ts
@@ -3,8 +3,8 @@ const WEEK = 7 * 24 * 3600 * 1000;
 const options = (messageDate: Date): Record<string, string> => {
   const now = new Date();
   if (now.getDay() === messageDate.getDay()) return { hour: "numeric", minute: "numeric" };
-  if (now.valueOf() - messageDate.valueOf() < WEEK) return { weekday: "long" };
-  return { year: "numeric", month: "long", day: "numeric" };
+  if (now.valueOf() - messageDate.valueOf() < WEEK) return { weekday: "short" };
+  return { year: "numeric", month: "short", day: "numeric" };
 };
 export function formatMessageTime(time: string): string {
   const messageDate = new Date(time);
